+++ AlaktuManager/Server/Repository/IRepository.cs
@@ -5,10 +5,10 @@ namespace AlaktuManager.Server.Repository
     public interface IRepository<T> where T : class, IEntity
     {
         Task<List<T>> GetAll();
-        Task<T> Get(int id);
+        Task<T> Get(Int64 id);
         Task<T> Add(T entity);
         Task<T> Update(T entity);
-        Task<T> Delete(int id);
+        Task<T> Delete(Int64 id);
         Task<List<T>> GetView();
     }
 }
